Time: 90 ms (25.58%), Space: 19 MB (32.79%) - LeetHub

+++ 1473-find-the-longest-substring-containing-vowels-in-even-counts/1473-find-the-longest-substring-containing-vowels-in-even-counts.cpp
@@ -22,7 +22,7 @@ public:
         {
             //ab maai bit mask tabhi change karunga agr vowel hai kyuki maine poora 26 bits ka mask leliya
             if(s[i]=='a' || s[i]=='e' || s[i]=='i' || s[i]=='o' || s[i]=='u')
-                count=count^(1<<((int)(s[i]-'a')));
+                count=count^(1<<((s[i]-'a')));
             if(mp.count(count))//agr uska count 0 se jaada hai toh dekha hua hai usko mtlb
             {
                 ans=max(ans,(i-mp[count]));
